feat(finance): modularize quick transfer modal into dedicated component
- Added `finance-quick-transfer-modal.tsx` as a new file to encapsulate the Quick Transfer Modal functionality.
- Implemented state management, transfer logic, and UI rendering in the new component.
- Removed the inlined modal logic from `finance-quick-transfer.tsx` to improve modularity and readability.
- Updated `finance-quick-transfer.tsx` to utilize the new `FinanceQuickTransferModal` component.

This change simplifies the codebase by isolating the modal logic and maintains feature parity with the previous implementation.
